Make light series of Latin Modern Typewriter font the default
The default Latin Modern Typewriter font appears a bit close to the
bold series and so to make the distinction appear more pronounce,
the light series has been suggested to be used as the default.
From "An exploration of the Latin Modern fonts" by Will Robertson,
The PracTEX Journal, TPJ2006No01, 2006-02-01.

This font is used for command, and output verbatim texts.